// web/src/types/Publication.ts
export interface Url {
    name: string; // Name of the URL (e.g., "Arxiv")
    url: string; // The actual URL
    image?: string; // Optional image URL
}

export interface Publication {
    id?: string; // Optional, as it will be generated by MongoDB
    title: string; // Title of the publication
    authors: string; // Authors as a single string
    description: string; // Description of the publication
    urls: Url; // Nested object for URL details
    order: number; // Default is 0, used for sorting
    createdAt?: string; // Optional, added by Mongoose timestamps
    updatedAt?: string; // Optional, added by Mongoose timestamps
}